Addition of projects page
Added the part of the site that needs to contain 2-4 computer science projects that influenced me. These projects are stored in a 2x2 grid, and their associated images have been added, each in three different sizes using srcset to make them fully responsive. The background color for this page has been changed to a dark blue