Contact page and Inspiring Projects modifications
Modified the project descriptions for Overwatch and Serato so that they took more of a software development perspective to the description. The contact page has also been added and fully styled to include all of the required fields and validating them as well. The navbar now contains an option for the Contact page