Responsiveness update
The site is now responsive for mobile, tablet, and desktop. Mobile layout displays when screen goes smaller than 600px (Navbar displays hamburger menu, grid is one column), and tablet layout is triggered when screen is between 600 and 1000px (navbar displays on the side) , and anything over 1000px shows the desktop site. In the grid, extra paragraphs have been added that display specifically for the mobile layout due to the grid dimensions being different